Add nextjs types + fix helpers & styles

+++ src/Helpers/local-storage.ts
@@ -8,7 +8,16 @@ type LocalStorage = {
 
 type LocaleStorageKey = keyof LocalStorage
 
+const warn = (message: string) => {
+  console.log('Window is not defined', message)
+}
+
 export const getStoredItem = <K extends LocaleStorageKey> (key: K): LocalStorage[K] | undefined => {
+  if (window === undefined) {
+    warn(` - Trying to get key : ${key}`)
+    return undefined
+  }
+
   const value = window.localStorage.getItem(key)
 
   if (value !== null) {
@@ -22,13 +31,28 @@ export const getStoredItem = <K extends LocaleStorageKey> (key: K): LocalStorage
 }
 
 export const storeItem = <K extends LocaleStorageKey> (key: K, value: LocalStorage[K]) => {
+  if (window === undefined) {
+    warn(` - Trying to store key : ${key} with value : ${value}`)
+    return
+  }
+
   window.localStorage.setItem(key, JSON.stringify(value))
 }
 
 export const removeStoredItem = (key: LocaleStorageKey) => {
+  if (window === undefined) {
+    warn(` - Trying to remove key : ${key}`)
+    return
+  }
+
   window.localStorage.removeItem(key)
 }
 
 export const clearStore = () => {
+  if (window === undefined) {
+    warn(' - Trying to clear store')
+    return
+  }
+
   window.localStorage.clear()
 }
